+++ authomatic/providers/oauth1.py
@@ -1182,10 +1182,8 @@ class Yahoo(OAuth1):
 
     Supported :class:`.User` properties:
 
-    * birth_date
     * city
     * country
-    * gender
     * id
     * link
     * location
@@ -1195,6 +1193,8 @@ class Yahoo(OAuth1):
 
     Unsupported :class:`.User` properties:
 
+    * birth_date
+    * gender
     * locale
     * phone
     * postal_code
@@ -1204,10 +1204,8 @@ class Yahoo(OAuth1):
     """
 
     supported_user_attributes = core.SupportedUserAttributes(
-        birth_date=True,
         city=True,
         country=True,
-        gender=True,
         id=True,
         link=True,
         location=True,
@@ -1257,17 +1255,6 @@ class Yahoo(OAuth1):
             # probably user hasn't activated Yahoo Profile
             user.city = None
             user.country = None
-
-        _date = _user.get('birthdate')
-        _year = _user.get('birthYear')
-
-        if _date and _year:
-            _full = _date + '/' + _year
-            try:
-                user.birth_date = datetime.datetime.strptime(_full, "%m/%d/%Y")
-            except BaseException:
-                user.birth_date = _full
-
         return user